requirements.txt
duckduckgo-search
arxiv
google-generativeai
